+++ common/src/main/java/ctbrec/io/HttpClient.java
@@ -29,6 +29,8 @@ import okhttp3.OkHttpClient.Builder;
 import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.Route;
+import okhttp3.WebSocket;
+import okhttp3.WebSocketListener;
 
 public abstract class HttpClient {
     private static final transient Logger LOG = LoggerFactory.getLogger(HttpClient.class);
@@ -218,5 +220,10 @@ public abstract class HttpClient {
     public void logout() {
         getCookieJar().clear();
         loggedIn = false;
+    }
+
+    public WebSocket newWebSocket(String url, WebSocketListener l) {
+        Request request = new Request.Builder().url(url).build();
+        return client.newWebSocket(request, l);
     }
 }